Stable after debugging over 1600 experiment case. Numerical errorrs on MOSEK should be fixed, hopefully

+++ script/create.py
@@ -9,21 +9,23 @@ import os
 directory = '../test/portfolio-output/'
 files = os.listdir(directory)
 for efile in files:
-    os.remove(directory + efile)
+    if ".txt" in efile:
+        os.remove(directory + efile)
 
 
 
 # Step 0: Condor for AA problems
 
 # Step 0.0: Arrays
-capital = [10000]
-ret = [0.01, 0.04]
-cardinality = [5, 10]
+capital = [100000]
+ret = [0.041, 0.061]
+cardinality = [1, 3, 5]
 branch = ['mf','hc'] # random is out
 cut = ['mf', 'hc'] # random is out
-search = ['df0', 'df1', 'breadth', 'best']
-cutiter = [1]
-cutperiter = [1]
+search = ['df0', 'df1', 'best'] # breadth is out
+cutiter = [1, 2]
+cutperiter = [1, 3]
+cutlim = [10, 100]
 
 # Step 0.1: Open file
 cfile = open('condor.sub','w')
@@ -34,7 +36,8 @@ common_command = \
 'Executable = ../test/portfolio \n\
 Universe   = vanilla\n\
 getenv     = true\n\
-transfer_executable = false \n\n'
+transfer_executable = false \n\
+\n\n'
 cfile.write(common_command)
 
 # Step 0.3: Loop over arrays
@@ -46,35 +49,43 @@ for (c,r,br,cu,sea) in comb:
     index = index + 1
     ci = 0
     cp = 0
+    cl = 0
     # run command
     run_command =  \
-'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -C %d -r %f -f 0 -o 2\n\
-output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, c, r, index)
+'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 0 -l %d -i %d -p %d -C %d -r %f -f 0 -o 2\n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, cl, ci, cp, c, r, index, index)
     # write to file
     cfile.write(run_command)
-    prob_index = '%04d - round\tb:%s c:%s s:%s i:%d p:%d C:%d r:%.2f\n' %(index, br, cu, sea, ci, cp, c, r)
+    prob_index = '%04d\tround\tb:%s\tc:%s\ts:%s\tl:%d\ti:%d\tp:%d\tC:%d\tr:%.4f\tleader\n' %(index, br, cu, sea, cl, ci, cp, c, r)
     ifile.write(prob_index)
 
     # x = 1
     index += 1
+    cl = 100
     run_command =  \
-'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 1 -l 100 -i %d -p %d -C %d -r %f -f 0 -o 2\n\
-output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, c, r, index)
+'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 1 -l %d -i %d -p %d -C %d -r %f -f 0 -o 2\n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, cl, ci, cp, c, r, index, index)
     # write to file
     cfile.write(run_command)
-    prob_index = '%04d - round\tb:%s c:%s s:%s i:%d p:%d C:%d r:%.2f x:1\n' %(index, br, cu, sea, ci, cp, c, r)
+    prob_index = '%04d\tround\tb:%s\tc:%s\ts:%s\tl:%d\ti:%d\tp:%d\tC:%d\tr:%.4f\tx:1\n' %(index, br, cu, sea, cl, ci, cp, c, r)
     ifile.write(prob_index)
     
-    comb2 = [(ci,cp) for ci in cutiter for cp in cutperiter]
-    for (ci,cp) in comb2:
+    comb2 = [(ci,cp,cl) for ci in cutiter for cp in cutperiter for cl in cutlim]
+    for (ci,cp,cl) in comb2:
         # run command
         index += 1
         run_command =  \
-'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -C %d -r %f -f 0 -o 2\n\
-output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, c, r, index)
+'arguments  = -type roundlot -d AA -a 20 -b %s -c %s -s %s -x 0 -l %d -i %d -p %d -C %d -r %f -f 0 -o 2\n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, cl, ci, cp, c, r, index, index)
         # write to file
         cfile.write(run_command)
-        prob_index = '%04d - round\tb:%s c:%s s:%s i:%d p:%d C:%d r:%.2f\n' %(index, br, cu, sea, ci, cp, c, r)
+        prob_index = '%04d\tround\tb:%s\tc:%s\ts:%s\tl:%d\ti:%d\tp:%d\tC:%d\tr:%.4f\n' %(index, br, cu, sea, cl, ci, cp, c, r)
         ifile.write(prob_index)
 
 #  -----------  CARDINALITY -----------
@@ -86,28 +97,34 @@ for (r,br,cu,sea,k) in comb3:
     cp = 0
     run_command =  \
     'arguments  = -type cardinality -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
-    output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index)
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
     # write to file
     cfile.write(run_command)
-    prob_index = '%04d - cardi\tb:%s c:%s s:%s i:%d p:%d k:%d r:%.2f\n' %(index, br, cu, sea, ci, cp, k, r)
+    prob_index = '%04d\tcardi\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tleader\n' %(index, br, cu, sea, ci, cp, k, r)
     ifile.write(prob_index)
     
     index += 1
     run_command =  \
     'arguments  = -type cardinality -d AA -a 20 -b %s -c %s -s %s -x 1 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
-    output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index)
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index,index)
     # write to file
     cfile.write(run_command)
-    prob_index = '%04d - cardi\tb:%s c:%s s:%s i:%d p:%d k:%d r:%.2f x:1\n' %(index, br, cu, sea, ci, cp, k, r)
+    prob_index = '%04d\tcardi\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tx:1\n' %(index, br, cu, sea, ci, cp, k, r)
     ifile.write(prob_index)
     
     index += 1
     run_command =  \
     'arguments  = -type cardinality -d AA -a 20 -b %s -c %s -s %s -x 2 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
-    output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index)
+    output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
     # write to file
     cfile.write(run_command)
-    prob_index = '%04d - cardi\tb:%s c:%s s:%s i:%d p:%d k:%d r:%.2f x:2\n' %(index, br, cu, sea, ci, cp, k, r)
+    prob_index = '%04d\tcardi\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tx:2\n' %(index, br, cu, sea, ci, cp, k, r)
     ifile.write(prob_index)
     
     comb2 = [(ci,cp) for ci in cutiter for cp in cutperiter]
@@ -116,11 +133,70 @@ for (r,br,cu,sea,k) in comb3:
         index += 1
         run_command =  \
             'arguments  = -type cardinality -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
-output     = ../test/portfolio-output/out.%04d.txt\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index)
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
         # write to file
         cfile.write(run_command)
-        prob_index = '%04d - cardi\tb:%s c:%s s:%s i:%d p:%d k:%d r:%.2f\n' %(index, br, cu, sea, ci, cp, k, r)
+        prob_index = '%04d\tcardi\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\n' %(index, br, cu, sea, ci, cp, k, r)
         ifile.write(prob_index)
+
+
+#  -----------  SINGLE CARDINALITY -----------
+
+#comb3 = [(r,br,cu,sea,k) for r in ret for br in branch for cu in cut for sea in search for k in cardinality]
+for (r,br,cu,sea,k) in comb3:
+    index = index+1
+    ci = 0
+    cp = 0
+    run_command =  \
+    'arguments  = -type single -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
+    # write to file
+    cfile.write(run_command)
+    prob_index = '%04d\tsingle\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tleader\n' %(index, br, cu, sea, ci, cp, k, r)
+    ifile.write(prob_index)
+    
+    index += 1
+    run_command =  \
+    'arguments  = -type single -d AA -a 20 -b %s -c %s -s %s -x 1 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index,index)
+    # write to file
+    cfile.write(run_command)
+    prob_index = '%04d\tsingle\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tx:1\n' %(index, br, cu, sea, ci, cp, k, r)
+    ifile.write(prob_index)
+    
+    index += 1
+    run_command =  \
+    'arguments  = -type single -d AA -a 20 -b %s -c %s -s %s -x 2 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
+    output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
+    # write to file
+    cfile.write(run_command)
+    prob_index = '%04d\tsingle\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\tx:2\n' %(index, br, cu, sea, ci, cp, k, r)
+    ifile.write(prob_index)
+    
+    comb2 = [(ci,cp) for ci in cutiter for cp in cutperiter]
+    for (ci,cp) in comb2:
+        # run command
+        index += 1
+        run_command =  \
+            'arguments  = -type single -d AA -a 20 -b %s -c %s -s %s -x 0 -l 100 -i %d -p %d -r %f -f 0 -o 2 -k %d -ct quadratic \n\
+output     = ../test/portfolio-output/out.%04d.txt\
+\nlog      = ../test/portfolio-output/log/%04d.txt\
+\nqueue 1\n\n' %(br, cu, sea, ci, cp, r, k, index, index)
+        # write to file
+        cfile.write(run_command)
+        prob_index = '%04d\tsingle\tb:%s\tc:%s\ts:%s\tl:100\ti:%d\tp:%d\tk:%d\tr:%.4f\n' %(index, br, cu, sea, ci, cp, k, r)
+        ifile.write(prob_index)
+
+
+
